Use common dasboot generation for compile and run
This fixes #379.
