+++ nin/backend/compile.js
@@ -11,6 +11,7 @@ const rmdir = require('rimraf');
 const shaderGen = require('./shadergen').shaderGen;
 const stream = require('stream');
 const utils = require('./utils');
+const dasbootGen = require('./dasbootgen');
 const walk = require('walk');
 
 
@@ -164,34 +165,34 @@ const compile = function(projectPath, options) {
         fs.writeFileSync(projectPath + '/gen/files.js', new Buffer(data));
         projectSettings.generate(projectPath);
         shaderGen(projectPath, function() {
-          process.stdout.write(chalk.yellow('\nRunning closure compiler'));
-          const globPaths = [
-            __dirname + '/../dasBoot/lib/*.js',
-            __dirname + '/../dasBoot/*.js',
-            projectPath + '/lib/*.js ',
-            projectPath + '/gen/*.js ',
-            projectPath + '/src/*.js',
-          ];
-          const jsCode = [].concat.apply(
-            [], globPaths.map(globPath => glob.sync(globPath)))
-            .map(path => ({
-              src: fs.readFileSync(path, 'utf8'),
-              path
-            }));
-          const out = closureCompiler.compile({
-            jsCode: jsCode
+          dasbootGen(projectPath, function() {
+            process.stdout.write(chalk.yellow('\nRunning closure compiler'));
+            const globPaths = [
+              projectPath + '/gen/*.js ',
+              projectPath + '/lib/*.js ',
+              projectPath + '/src/*.js',
+            ];
+            const jsCode = [].concat.apply(
+              [], globPaths.map(globPath => glob.sync(globPath)))
+              .map(path => ({
+                src: fs.readFileSync(path, 'utf8'),
+                path
+              }));
+            const out = closureCompiler.compile({
+              jsCode: jsCode
+            });
+            if(out.errors.length) {
+              renderError();
+              out.errors.map(console.error);
+              process.exit(1);
+            } else if(out.warnings.length) {
+              renderWarn();
+              out.warnings.map(console.error);
+            } else {
+              renderOK();
+            }
+            collect(out.compiledCode);
           });
-          if(out.errors.length) {
-            renderError();
-            out.errors.map(console.error);
-            process.exit(1);
-          } else if(out.warnings.length) {
-            renderWarn();
-            out.warnings.map(console.error);
-          } else {
-            renderOK();
-          }
-          collect(out.compiledCode);
         });
       });
     });
